+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Dev
 
+- Add `Cmp.min_by/2` and `Cmp.max_by/2`
+
 ## v0.1.1 (2023-03-05)
 
 - Fix broken link in doc